cart/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-bootstrap5==24.2
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==26.0.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
